+++ sentry-samples/sentry-samples-spring-boot-jakarta-opentelemetry-noagent/src/main/java/io/sentry/samples/spring/boot/jakarta/SentryDemoApplication.java
@@ -1,5 +1,7 @@
 package io.sentry.samples.spring.boot.jakarta;
 
+import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.api.trace.Tracer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -59,4 +61,16 @@ public class SentryDemoApplication {
   //    trigger.setCronExpression("0 0/5 * ? * *"); // every five minutes
   //    return trigger;
   //  }
+
+  //  @Bean
+  //  public Tracer getTracer(OpenTelemetry openTelemetry) {
+  //    GlobalOpenTelemetry.set(openTelemetry);
+  //    return openTelemetry.getTracer("tracerForSpringBootDemo");
+  ////    return GlobalOpenTelemetry.getTracer("tracerForSpringBootDemo");
+  //  }
+
+  @Bean
+  public Tracer tracer(OpenTelemetry openTelemetry) {
+    return openTelemetry.getTracer("tracerForSpringBootDemo");
+  }
 }