+++ sentry-spring-jakarta/src/main/java/io/sentry/spring/jakarta/opentelemetry/SentryOpenTelemetryAgentWithoutAutoInitConfiguration.java
@@ -29,7 +29,7 @@ public class SentryOpenTelemetryAgentWithoutAutoInitConfiguration {
     return options -> {
       SentryIntegrationPackageStorage.getInstance()
           .addIntegration("SpringBoot3OpenTelemetryAgentWithoutAutoInit");
-      OpenTelemetryUtil.applyOpenTelemetryOptions(options);
+      OpenTelemetryUtil.applyOpenTelemetryOptions(options, true);
     };
   }
 }