+++ sentry-spring-jakarta/src/main/java/io/sentry/spring/jakarta/opentelemetry/SentryOpenTelemetryNoAgentConfiguration.java
@@ -32,7 +32,7 @@ public class SentryOpenTelemetryNoAgentConfiguration {
       SentryIntegrationPackageStorage.getInstance()
           .addIntegration("SpringBoot3OpenTelemetryNoAgent");
       SentryAutoConfigurationCustomizerProvider.skipInit = true;
-      OpenTelemetryUtil.applyOpenTelemetryOptions(options);
+      OpenTelemetryUtil.applyOpenTelemetryOptions(options, false);
     };
   }
 }